Fix uint32_t to float conversion warning for Linux GCC compatibility
- Use explicit static_cast for random_seed to float conversion
- Ensures clean build on Linux with strict GCC warnings

+++ include/rtpghi/rtpghi.hpp
@@ -437,7 +437,7 @@ namespace rtpghi
             {
                 random_seed =
                     (random_seed * constants::LCG_MULTIPLIER + constants::LCG_INCREMENT) & constants::LCG_MASK;
-                output.phases[m] = (random_seed / float(constants::LCG_MASK)) * constants::TWO_PI;
+                output.phases[m] = (static_cast<float>(random_seed) / static_cast<float>(constants::LCG_MASK)) * constants::TWO_PI;
             }
         }
